Proiect/requirements.txt
Flask==1.1.1
requests==2.23.0
azure-ai-textanalytics==1.0.0b4
azure-cognitiveservices-speech==1.11.0
GoogleAppEngineCloudStorageClient==1.9.22.1
google-cloud-storage==1.26.0
google-cloud-vision==1.0.0
google-cloud-translate==2.0.1
google-cloud-datastore==1.11.0
google-auth==1.11.2
google-cloud-speech==1.3.2